+++ semantic-conventions/src/tests/semconv/templating/test_markdown.py
@@ -123,6 +123,16 @@ class TestCorrectMarkdown(unittest.TestCase):
     def test_event_renamed(self):
         self.check("markdown/event_renamed/")
 
+    def test_metric_tables(self):
+        self.check(
+            "markdown/metrics_tables",
+            extra_yaml_files=[
+                "yaml/general.yaml",
+                "yaml/http.yaml",
+                "yaml/metrics.yaml",
+            ],
+        )
+
     def testSamplingRelevant(self):
         self.check("markdown/sampling_relevant/")
 
@@ -139,6 +149,7 @@ class TestCorrectMarkdown(unittest.TestCase):
         *,
         expected_name="expected.md",
         extra_yaml_dirs: Sequence[str] = (),
+        extra_yaml_files: Sequence[str] = (),
         assert_raises=None
     ) -> Optional[BaseException]:
         dirpath = Path(self.get_file_path(input_dir))
@@ -154,6 +165,9 @@ class TestCorrectMarkdown(unittest.TestCase):
             for fname in Path(self.get_file_path(extra_dir)).glob("*.yaml"):
                 print("Parsing", fname)
                 semconv.parse(fname)
+        for fname in map(self.get_file_path, extra_yaml_files):
+            print("Parsing ", fname)
+            semconv.parse(fname)
 
         semconv.finish()